dev: refactor tasks.py into maintainable modules
Fixes #501
