+++ automation/error_handling.py
@@ -1,0 +1,9 @@
+def exit_if_error_in_stdout(result: Result):
+    # Find N remaining using regex
+
+    if "error" in result.stdout:
+        errors_remaining = re.findall(r"\d+(?=( remaining))", result.stdout)[
+            0
+        ]  # testing
+        if errors_remaining != "0":
+            exit(0)
